Get rid of runs.pre hook
