+++ action.yml
@@ -31,5 +31,4 @@ outputs:
 
 runs:
   using: "node20"
-  pre: "install-deps.js"
   main: "dist/index.js"
